Consolidate roles
The counts work, but the alignment is a bit off. I'm not exactly sure how to fix it using display:flex, but wanted to at least get the consolidation logic in place. I tried incrementing the counts on the page objects themselves while they were being rendered, but it ended up being more trouble than it was worth doing it that way so I tacked them onto the end. Hope this helps! :)